Allow Table `function` to opt-in to return arbitrary combinations of [0, output.size()] x [HAVE_MORE_OUTPUT, FINISHED] (https://redirect.github.com/duckdb/duckdb/pull/19566)
Follow up to https://github.com/duckdb/duckdb/pull/19422.

File readers might have enough information to know a given amount of
rows needs to be returned, and those are the last ones.
Currently this means `{[...some rows...], HAVE_MORE_OUTPUT} + {[],
FINISHED}`, while this could be just a `{[... some rows ... ],
FINISHED}`.

Conversely `{[], HAVE_MORE_OUTPUT}` was currently forbidden, meaning a
scan would have to loop looking for at least a single row, that might
block for a while in case of very selective filters.

With this PR, when opting in away from the `results_execution_mode =
AsyncResultsExecutionMode::SYNCHRONOUS` then one can freely implicitly
return any of the following states:
* `{[...some rows], HAVE_MORE_OUTPUT}`
* `{[...some rows], FINISHED}`
* `{[], HAVE_MORE_OUTPUT}`
* `{[], FINISHED}`
* `{[], BLOCKED}`

While in the default, not opted in case, only 1st, 4th and 5th are
available.

Note that as per this PR this does change the behaviour of scans, but
not really in a visible way, since the file readers are not taking
advantage (yet).

Added testing code that checks both old mode (used in legacy code such
as some extensions) and new code works via a setting that allows to
change execution strategy.

Examples:
```
./build/release/test/unittest

Upstream-commit: https://github.com/duckdb/duckdb/commit/35616dc32aeab622dc79020caeb263d9dec21b37
--- expected to work
./build/release/test/unittest --on-init "SET debug_physical_table_scan_execution_strategy=TASK_EXECUTOR;"
--- equivalent to default, expected to work
./build/release/test/unittest --on-init "SET debug_physical_table_scan_execution_strategy=SYNCHRONOUS;"
--- equivalent to old default, expected to work, a configuration is added to test this case specifically in Main.yml
./build/release/test/unittest --on-init "SET debug_physical_table_scan_execution_strategy=TASK_EXECUTOR_BUT_FORCE_SYNC_CHECKS;" 
---- expected to NOT work in general, will fail in 200 some tests, that are the one where there is a change in behaviour
```