Increase cast-cost of old-style implicit cast to string (https://redirect.github.com/duckdb/duckdb/pull/19621)
This PR fixes https://github.com/duckdb/duckdb-python/issues/148

The issue is that `list_extract` now has two overloads, one for a
templated list `LIST<T>` and for concrete `VARCHAR` inputs. When binding
a function we add a really high cost to selecting a templated overload
to ensure we always pick something more specific if available. With our
current casting rules, we are unable to cast `VARHCAR[]` to `VARCHAR`,
and therefore fall back to the list-template as expected. But with
old-style casting rules we allow `VARCHAR[]` to `VARCHAR` by also adding
a high cost penalty, but its still lower than the cost of casting to the
template - even though that would be the better alternative.

With old-style casting we basically always have a lower-cost "fallback"
option than selecting a template overload. While we should overhaul our
casting system to evaluate the cast cost along more axes than just
"score", this PR fixes this specific case by just cranking up the cost
of old-style implicit to-string casts.

Upstream-commit: https://github.com/duckdb/duckdb/commit/a0a46d6ed06dd962a4d6eeb01f3e14f8b275cec4