[ART] Pretty Printer + Verify (https://redirect.github.com/duckdb/duckdb/pull/19558)
Separate ART tree VerifyAndToString into two separate functions.
Prettify the ART tree ToString method, as well as an ASCII mode flag for
displaying characters instead of byte values (For nested trees, it still
prints byte values since those are row IDs). This is similar to ToString
methods on chunks and expressions; there is no testing on this since it
is just a debugging nicety. Example output:

```
ART: 
Node4
a,
  Prefix: |p|
  Node4
  p,
    Prefix: |l|
    Node4
    e,
      Prefix: |0|
      Inlined Leaf [row ID: 0]
    i,
      Prefix: |c|a|t|
      Node4
      i,
        Prefix: |o|n|0|
        Inlined Leaf [row ID: 1]
      o,
        Prefix: |r|0|
        Inlined Leaf [row ID: 2]
    y,
      Prefix: |0|
      Inlined Leaf [row ID: 3]
  r,
    Prefix: |i|c|o|t|0|
    Inlined Leaf [row ID: 4]
b,
  Prefix: |a|n|
  Node4
  a,
    Prefix: |n|a|0|
    Inlined Leaf [row ID: 5]
  d,
    Node4
    0,
      Inlined Leaf [row ID: 6]
    a,
      Prefix: |n|a|0|
      Inlined Leaf [row ID: 7]
  n,
    Prefix: |e|r|0|
    Inlined Leaf [row ID: 8]
  q,
    Prefix: |u|e|t|0|
    Inlined Leaf [row ID: 9]
```

Upstream-commit: https://github.com/duckdb/duckdb/commit/b32f072ade4879a711c902d441fc7317a0dd5508