Encapsulate `BaseScalarFunction` properties (https://redirect.github.com/duckdb/duckdb/pull/19632)
This PR wraps the properties of the `BaseScalarFunction` class in
getters/setters, and updates all call-sites in duckdb, `core_functions`
and `parquet` extensions.

The actual underlying fields remain public for now though to not
immediately break all out-of-core extensions.

I have a lot of changes planned for scalar/aggregate functions in the
future, and the motivation to add these getter/setters is to make it
easier to change the internals while keeping the interface the same.

Upstream-commit: https://github.com/duckdb/duckdb/commit/2d7fd335e1b8c310a6d6f6733acc8769fa4653d2