Buffer-managed query results (https://redirect.github.com/duckdb/duckdb/pull/19437)
This PR implements buffer-managed query results.

## The Issue
Currently, query results are either streaming, or materialized. This PR
focuses on the latter. Materialized query results are fully
memory-resident, and the allocations do not go through the buffer
manager. This can cause DuckDB to OOM without throwing an exception,
potentially getting killed by the OS, or freezing the OS if you do this:
```sql
SELECT * FROM tpch_sf1000.lineitem;
```
On a machine with, e.g., 16 GB of RAM.

The reason that query results aren't buffer-managed right now, is due to
a lifetime issue. DuckDB is in-process, and we hand over query results
to, e.g., Python. After this, the DB can be closed, destroying the
buffer manager. If the query result would be buffer-managed at this
point, touching it in the slightest would cause a segfault, as it would
try to use the buffer manager, which no longer exists.

With this PR, we enable buffer-managed query results, and tie their
lifetime to that of the database. If the database is closed, you'll get
an exception when trying to access the query result.

## `ResultSetManager`
Instead of `QueryResult`s holding a `ColumnDataCollection`, they now
hold a `ManagedResultSet`. To access the `ColumnDataCollection` within,
you need to `Pin` the `ManagedResultSet`, obtaining a `PinnedResultSet`.
Als long as `PinnedResultSet` is within scope, the `DatabaseInstance`
cannot be closed, and the buffer-managed `ColumnDataCollection` will
stay available.

The `DatabaseInstance` now has a `ResultSetManager`, which has a
`weak_ptr<ManagedResultSet>` for each open query result. Once the
`DatabaseInstance` is closed, the `ColumnDataCollection`s within the
`ManagedResultSet`s will be destroyed. If a `ManagedResultSet` is
destroyed, it will be removed from the `ResultSetManager`.

Note that all of this is optional: in-memory query results are still
possible, and while `weak_ptr`s to these are also held by
`ResultSetManager`, their lifetime is not tied to that of the
`DatabaseInstance`.

## Example
Open the CLI with a persistent DB:
```sql
-- low limit and disable spilling
set memory_limit='100mb';
set temp_directory='';
-- create table larger than 100mb
create table range as from range(100_000_000);
-- this result does not fit in 100mb, but this query would succeed before, silently exceeding the limit
-- with this PR, an OOM exception is now thrown because we don't have a temp directory to spill it
from range;
-- enable spilling
set temp_directory='/tmp';
-- this now succeeds, the large result is spilled to the temp directory
from range;
```

## Future Work
I've currently only enabled this in the CLI and unit tester, as this PR
is already quite big, and I would prefer to add it to the CAPI and other
clients incrementally.

Upstream-commit: https://github.com/duckdb/duckdb/commit/ce5c1e7ef0335c9da860ae44c6a127f8a7d65309