CI Fuzzer runtime error (https://redirect.github.com/duckdb/duckdb/pull/19630)
Fixes: https://github.com/duckdb/duckdb-fuzzer/issues/4298 and
https://github.com/duckdb/duckdb-fuzzer/issues/4299

The runtime error occurred because we didn't check for NULL values
before attempting to cast them as booleans. This was happening in
several places, which is why it caused multiple issues. I've addressed
it everywhere (hopefully) to prevent future surprises. Additionally, the
query generated by the optimizer was somewhat confusing, so I simplified
it to the minimal version needed to reproduce the problem in my tests.

Upstream-commit: https://github.com/duckdb/duckdb/commit/0dd686e5c40328f86f6e39984140ba4461ecb0ec