Don't truncate test names to 80 characters (https://redirect.github.com/duckdb/duckdb/pull/19642)
We sometimes have test names that are longer than 80 characters. These
currently get truncated to 80 characters in the CI output and that can
make it quite hard to find out exactly what test failed (especially if
the test crashed instead of failing).

This changes the logic to only do the truncation for terminal windows
because that's where we care about it to make the `\r` trick work. In
passing this also changes a few other details:
- Use the terminal width as the limit instead of a hardcoded 80
characters
- Truncate the start of the string instead of the end
- Use `\n` instead of `\r` for non-terminals

Upstream-commit: https://github.com/duckdb/duckdb/commit/963000678ad9d9f68b6636415e6f3ff73a65087c