CLI - Add paging support for output (https://redirect.github.com/duckdb/duckdb/pull/19676)
Supersedes https://github.com/duckdb/duckdb/pull/19004 - thanks @tobwen
for the initial implementation. I've taken the liberty of merging this
into the recent CLI rework and finishing this up.

This PR introduces paging for query results using a configurable pager.
By default, the pager is set to `automatic`, which means the pager only
triggers when there are more than 50 rows in the query result OR the
query result has more than 5 columns.

A special exception is made for `.mode duckbox` - since that already
truncates query results based on `.maxrows`. If we are using duckbox
mode and `.maxrows` (default 40) is smaller than the pager row
threshold, the pager is also never used. As a result, the pager is not
triggered by default unless either (1) `.mode` is changed, or (2)
`.maxrows` is changed, or (3) `.pager` is changed.

The pager can be configured using the `.pager` command:

```sql
-- always use the pager
.pager on
-- never use the pager
.pager off
-- use the pager when either (a) the row count, or (b) the column count, exceeds the specified limits
.pager automatic
-- trigger the pager when there are more than 100 rows
.pager set_row_threshold 100
-- trigger the pager when there are more than 7 columns
.pager set_column_threshold 7
-- configure the pager to use
.pager "my_favorite_pager"
```

The default pager used is determined by the first value found, in-order:

1. Value in `DUCKDB_PAGER` environment variable
2. Value in `PAGER` environment variable
3. `less -RX` on Linux/MacOS, and `more` on Windows. 

To see the pager, you can use e.g.:

```
.mode box
FROM range(10000);
```

### MoreRowsThan

In order to facilitate implementation of the pager, we introduce a new
method to the QueryResult: `MoreRowsThan`. This method returns whether
or not the query result has more rows than a specified amount, without
consuming any rows, i.e. subsequent calls to `Fetch()` will still return
the rows that were checked.

For materialized query results this method is trivial (just comparing
the count in the collection). For streaming query results, we fetch
until either (1) the result is exhausted, or (2) we have fetched more
than `n` rows. Any chunks we fetch during this call are buffered in
`stored_chunks` and emitted during subsequent fetches.

In order to facilitate this from an API level - we move the actual
implementation of Fetch from `FetchRaw` to `FetchInternal`.

Upstream-commit: https://github.com/duckdb/duckdb/commit/abcf96328794d01625567a27a04e04d5794b6d94