`GEOMETRY` Rework: Part 4 - Fixup Parquet Extension + Add Arrow Support (https://redirect.github.com/duckdb/duckdb/pull/19476)
This is a followup PR that builds on top of
https://github.com/duckdb/duckdb/pull/19439. Please have a look at
https://github.com/duckdb/duckdb/pull/19136 for the context behind this
PR.

This PR fixes up the remaining issues in the parquet extension related
to geometries. When reading geometry columns we now push an expression
column reader on top of the underlying blob column reader to perform the
WKB parsing with `ST_GeomFromWKB`. `ST_GeomFromWKB` now actually checks
that the input is valid WKB and also converts from big-endian WKB to
little-endian If required. This can be optimized further, but It's good
enough for now.

I've also added support for converting geometry columns to/from arrow
arrays with geoarrow extension metadata. This code is basically lifted
[straight from the spatial
extension](https://github.com/duckdb/duckdb-spatial/blob/v1.4-andium/src/spatial/spatial_geoarrow.cpp).

Upstream-commit: https://github.com/duckdb/duckdb/commit/b3874789f7d7a516ce98048f23260b4a35dd0a1c