SegmentTree: move `next` and `index` out of SegmentBase and into SegmentNode (https://redirect.github.com/duckdb/duckdb/pull/19686)
This PR reworks the segment tree class so that the actual traversal of
nodes within the segment tree happens fully within the `SegmentNode` -
and is not mixed between the segment tree and the `SegmentBase`
(`RowGroup` or `ColumnSegment`). This isolates the logic for traversing
the segment tree and decouples the `RowGroup` from the segment tree
traversal.

This doesn't accomplish much in and of itself - but a follow-up PR based
on this changeset can allow row groups to be stored in multiple
different segment trees simultaneously. This could then allow
checkpoints to be done in a lock-free manner while threads by
constructing a new segment tree, while other threads could still be
scanning the old segment tree.

Upstream-commit: https://github.com/duckdb/duckdb/commit/b67f32137c441081473b0ff15e2bfa1bb5d1b87d