Add `Clear` method to `BaseAppender` (https://redirect.github.com/duckdb/duckdb/pull/19623)
## Summary
This Pull Request adds a new `Clear` function to the `Appender`,
enabling the clearing of pending data without flushing.

### Motivation
When closing an `Appender`, a `Flush` will also be triggered.
However, sometimes, we wouldn't want to flush the data, for instance,
when there is a problem with the flushing.
In such case, we would sometimes prefer to only clear the appender and
close it.
We enable this capability by exposing the `Clear` method.

### Changes
- `Clear` method
- Added a test for verifying the `Clear` method correctly clears data
and that the next `Flush` won't write the data that should have been
cleared.
- Added a test that simulates a Heavy `QueryAppender` that should be
interrupted and then cleared, before being closed, to ensure the `Clear`
method actually "enables" the `Close` method for such scenarios.

### Upcoming Next
A PR for exposing the `Clear` method through the C-API.

Upstream-commit: https://github.com/duckdb/duckdb/commit/237b7f76cad3820e347a5d623567cbff9bf23b8b