`CommonSubplan` improvements (https://redirect.github.com/duckdb/duckdb/pull/19710)
This PR mostly reorganizes and cleans up the code for the
`CommonSubplan` optimizer.

While doing this, I also improved it slightly by allowing it to
materialize common subplans that do not end in a aggregate, distinct, or
window. The heuristic for this is to also materialize common subplans
that involve multiple base tables, where the estimated cardinality is
less than half of the maximum estimated cardinality of said base tables.
This improves TPC-DS Q27 by 3x. To make it more robust against join
ordering, I've moved the optimizer to run after the
`BuildProbeSideOptimizer`.

Upstream-commit: https://github.com/duckdb/duckdb/commit/23d01e6975f847a9f143e2f153304bdf06cf3803