implement proper equals for table functions (https://redirect.github.com/duckdb/duckdb/pull/19722)
Closes https://github.com/duckdb/duckdb/issues/19645

The issue is that during pipeline verification we compare all the
operators across pipelines and ensure that their equality is reflexive
(which is kind of a strange check actually), but in the case of table
scans we only compare the normal table function callback pointer. In the
MRE in the linked issue, the pipeline contains a UNNEST and a RANGE,
which are both table-in-out functions, but not _normal_ table functions,
so the function pointer(s) are null on both sides and they are assumed
to be equal.

I've fixed it by actually implementing the `==` and `!=` operates for
`TableFunction`s that compare all the properties, similar to how its
done for `ScalarFunction`.

Upstream-commit: https://github.com/duckdb/duckdb/commit/0efe5ccb5ba436fed2269d9b063995cc3990eaa2