fix bug in generate_grammar script (https://redirect.github.com/duckdb/duckdb/pull/19751)
Hi everyone,

there is a small bug in the `generate_grammar.py` script. 

In the python script, keywords are sorted being uppercase. However
bison/c++ seems to work with lowercase sorting. Since `[A-Z] < _ <
[a-z]` that can result in different "sorting" or looking for keywords
containing underscores. With the wrong order enforced by the script, the
parser is actually never able to match keywords that contain an
underscore. (I realised this when working on the `MATCH_RECOGNIZE`
grammar.)

It's a corner case since keywords with underscore are extremely rare.
But an easy fix. :)

Upstream-commit: https://github.com/duckdb/duckdb/commit/cf3860f589702891a356ec62fd845e9bfd114539