Fix initialization when all values are `NULL` in `SMALLER_BINARY` aggregation  (https://redirect.github.com/duckdb/duckdb/pull/19755)
Fixes: https://github.com/duckdblabs/duckdb-internal/issues/6289

When using `arg_max_nulls_last`, queries on tables where all values are
`NULL` returned a large integer (probably a placeholder as it is the
same as`INT_MAX`) instead of `NULL`.

This happens because in the aggregation phase the state was being set as
initialized too early, even before any valid (non-`NULL`) values were
encountered. Later, `Finalize()` assumes the state has a valid,
initialized value and tries to decode it , producing the large number.
If all input values are `NULL`, the state never gets properly
initialized.

The fix is to only set the state as initialized when there is at least
one non-NULL value.

Upstream-commit: https://github.com/duckdb/duckdb/commit/936b6435b1dcc20ca12aa64a392148396690deab