[C-API] Add support for catalog entry lookup (https://redirect.github.com/duckdb/duckdb/pull/19684)
This PR adds support for looking up catalog entries through the C-API. 

This is mostly useful for advanced extensions that need to inspect the
state of the catalog in the current transaction to e.g. detect if a
table already exists or to access state in a storage-extension catalog.

The original plan was to add support for storage extensions, but I
realized we probably need this first.

Upstream-commit: https://github.com/duckdb/duckdb/commit/737209ed29ab66052e559c67b89018062720a600