Storage: Remove `start` from RowGroup, ColumnData and ColumnSegment, and rework the way that `start` is defined in the `ColumnSegmentTree` (https://redirect.github.com/duckdb/duckdb/pull/19761)
Follow-up from https://github.com/duckdb/duckdb/pull/19686

Currently the `start` value for row groups is duplicated in many places
in the code. This (1) has lead to bugs before when these became
out-of-sync, and (2) introduces strict coupling between the RowGroup,
the columns within the row group, and the RowGroupSegmentTree, as the
start value is duplicated in many places within them.

For reference, here is what the current situation looks like for a
RowGroup at position `122800` (generally the location of the second
RowGroup with the default row group size). We can see that the `start`
value is duplicated in many places.

```
RowGroupSegmentTree
	...
	row_start = 122800
	RowGroup
		start = 122800
		StandardColumnData
			start = 122800
			ColumnSegmentTree
				SegmentNode
					row_start = 122800
					ColumnSegment
						start = 122800
			ValidityColumnData
				start = 122800
				ColumnSegmentTree
					SegmentNode
						row_start = 122800
						ColumnSegment
							start = 122800

```

In this PR we improve on this in two ways:

* We get rid of the duplicate `start` from RowGroup, ColumnData and
ColumnSegment
* We rework the `ColumnSegmentTree` so that the `start` position is
always 0, and not equal to the RowGroup it belongs to

i.e. this is now the situation:


```
RowGroupSegmentTree
	...
	row_start = 122800
	RowGroup
		StandardColumnData
			ColumnSegmentTree
				SegmentNode
					row_start = 0
					ColumnSegment
			ValidityColumnData
				ColumnSegmentTree
					SegmentNode
						row_start = 0
						ColumnSegment
```


This fully decouples the `RowGroup` from the segment tree, and also
decouples the `ColumnData` from the row group it belongs to. This allows
for much easier re-use of row groups, and also allows for row groups to
be part of multiple segment trees.

Upstream-commit: https://github.com/duckdb/duckdb/commit/20696d805ee421264950657119078ca621c8839b