Fix major big-endian issues (https://redirect.github.com/duckdb/duckdb/pull/19748)
I'm aware that big-endian platforms are officially not supported, but as
far as I see from the
[previous](https://github.com/duckdb/duckdb/issues/5548)
[discussions](https://github.com/duckdb/duckdb/issues/9714), patches are
welcome.

The PR fixes major big-endian related issues, which include the areas:
- sorting and ART indices;
- numeric conversion in C API;
- arrow UUID conversion;
- bit-packing;
- BIT conversion.

The intention was to do it in the least invasive way possible, even if
it means slightly suboptimal logic for big-endian path.
Please let me know if you want me to split it to several PRs.

The PR doesn't fix known big-endian issues with these:
- GEOMETRY;
- storage file portability;
- parquet;
- optional extensions.

If the PR is merged, they will be fixed in the next one.

Related issues: https://redirect.github.com/duckdb/duckdb/pull/9714, https://redirect.github.com/duckdb/duckdb/pull/5548.

Upstream-commit: https://github.com/duckdb/duckdb/commit/644effff469c7c698823ca182d6e66eb0ad8a2d2