PositionalScan: handle also HAVE_MORE_OUTPUT + empty chunk via iteration (https://redirect.github.com/duckdb/duckdb/pull/19824)
This is (in v1.4-andium) a case might only hit with weird plans, but
still a bug.
This has been uncovered in the `main` branch, where this becomes a more
common cases (and so more stress tested)

Note: there are still cases not properly handled, such as BLOCKED or
FINISHED + non-empty chunk, but those are to be fixed separately.

Upstream-commit: https://github.com/duckdb/duckdb/commit/8a7bc617e53a181dea0457d5b259a3abd70f54e7