AttachDatabase: first set whether file is remote, then do the rest (https://redirect.github.com/duckdb/duckdb/pull/19826)
Shuffle code around, while avoiding behavioural changes but for the fact
that we should go into `InsertDatabasePath` with the `access_mode`
properly set.

This avoids that when `access_mode` is `AUTOMATIC`, it's stored in the
Database cache without having considered that remote files needs to be
`READ_ONLY`.

This causes issues such as the added testcase, where adding the same
remote file twice (with different names) is invalid in automatic mode
(but valid if `READ_ONLY` is explicit)

Upstream-commit: https://github.com/duckdb/duckdb/commit/ff2930fb3061c7a92dadcffd4eb36d543c0bc955