Fix positional scan and restore task executor (https://redirect.github.com/duckdb/duckdb/pull/19825)
Cherry-pick https://github.com/duckdb/duckdb/pull/19824, to be reviewed
first (filed vs `v1.4-andium`, now cherry-picked on `main` as well),
that allows to the revert https://github.com/duckdb/duckdb/pull/19778.

Combined, this fixes a pre-existing bug and re-enable the async-of-sort
behaviour, so unblocking more work in that area.

Upstream-commit: https://github.com/duckdb/duckdb/commit/85590176796be8450fda62b33621a67f431e5c1e