`GEOMETRY` Rework: Part 4 - Step 2 - Adjust WKB conversion functions (https://redirect.github.com/duckdb/duckdb/pull/19848)
This is a followup PR that builds on top of
https://github.com/duckdb/duckdb/pull/19476. Please have a look at
https://github.com/duckdb/duckdb/pull/19136 for the context behind this
PR.

I realized I the `Geometry::FromBinary`/`Geometry::ToBinary` helper
functions need to be adjusted slightly so that they can be used to
implement the cast functions provided in `duckdb-spatial`. These casts
may move to core eventually, but for now this is required to integrate
the spatial extension with the new geometry type smoothly.

Upstream-commit: https://github.com/duckdb/duckdb/commit/d6dce8ba572b2fe603a0ed5ce4ae7076b5c07259