Enable reads to run concurrently while another connection is checkpointing (https://redirect.github.com/duckdb/duckdb/pull/19867)
Fixes https://github.com/duckdb/duckdb/issues/19500
Follow-up from https://github.com/duckdb/duckdb/pull/19843

This PR enables concurrent reads while checkpoints are running by
removing the table-level lock that used to be grabbed in
`TableScanInitGlobal`. This effectively prevents checkpoints from
blocking reads and vice versa. Checkpoints can start while readers are
active, and readers can start while a checkpoint is active.

This PR is relatively simple as most of the work has been done in the
previous chain of refactoring PRs
(https://github.com/duckdb/duckdb/pull/19686,
https://github.com/duckdb/duckdb/pull/19761,
https://github.com/duckdb/duckdb/pull/19843).

The crux of why this is now possible is that `Checkpoint` no longer does
an in-place modification of the `RowGroupSegmentTree` that is part of
the `RowGroupCollection`. Instead, it creates a new
`RowGroupSegmentTree` with (1) new `RowGroup` / `ColumnData` for any
entries that are checkpointed, and (2) direct references to old
(unchanged) entries. Only in the final stage, when the entire checkpoint
of a table is finished, do we update the `RowGroupSegmentTree` that is
referenced in the `RowGroupCollection`.

On the read side, when initiating a read, we keep a
`shared_ptr<RowGroupSegmentTree>` for the current segment tree we are
scanning. We then scan that segment tree for the remainder of the scan.
This ensures we can keep on scanning safely, as the
`RowGroupSegmentTree` we are scanning is kept alive by our scan state
even if it is replaced by the `Checkpoint`.

There are theoretically some more complications to this process - namely
if we were to do a `CHECKPOINT` twice old blocks could become
overwritten which would cause old `RowGroupSegmentTrees` to no longer be
valid, as they would point to blocks that have been overwritten with new
data. However, it is not possible to hit this issue currently because we
still have very aggressive locking around writes and checkpoints - it's
not actually possible to start a second `CHECKPOINT` while a read of old
data is still active in the current setup. However, this is an issue
that will need to be dealt with as we relax the write locks around
checkpoints in future PRs.

### Changes in this PR

Outside of removing the locks, this PR makes the following (relatively
minor) changes:

* Finish up work in `RowGroupCollection::Checkpoint` by making sure the
vacuum does not alter the original segment tree either
* Refactor SegmentNode<T> so that its members are private and there are
getters and setters
* Adds a lock to ART::SerializeToDisk so that we can safely write the
ART while reads are happening

Upstream-commit: https://github.com/duckdb/duckdb/commit/1f90f7d393fb7b68620d670fd5fd2efafa23a599