CLI: Enable support for mouse clicks in Linenoise - pressing Ctrl+Q allows a single mouse-click to be consumed to change the cursor position (https://redirect.github.com/duckdb/duckdb/pull/19869)
This PR enables mouse click support for changing the cursor position
through the `Ctrl+Q` shortcut. After pressing `Ctrl+Q` - a single mouse
event is consumed. If this event is a left click, the cursor is moved to
the position where the mouse is clicked. This can be useful for
navigating around a large query without having to repeatedly use the
arrow keys.

Unfortunately this is more or less the best we can do with mouse support
in the Linux / MacOS CLI due to the way mouse support works there.
Effectively mouse events are kind of "all or nothing" - either you
ignore mouse events, or you consume all mouse events, and all other
mouse behavior in the terminal is disabled - i.e. you can't scroll
anymore, select text anymore, etc.

I think this is a decent compromise. Consuming a single mouse event and
then reverting also prevents users from accidentally getting stuck in
this mode and getting confused. Feedback is welcome/appreciated.


This PR only enables this for Linux. Interestingly enough the Windows
terminal input support is much better and we could provide much better
mouse support there in theory - but we'll leave that for a future
iteration.

Upstream-commit: https://github.com/duckdb/duckdb/commit/6bb5529a59676f06f128dc48cf92d18c7980a0ca