[Compression] Fix an existing and a new bug in Roaring compression (https://redirect.github.com/duckdb/duckdb/pull/19878)
This PR fixes https://redirect.github.com/duckdb/duckdb/pull/19875

During flushing of the segment we use `AlignValue` before writing the
metadata, because it has to be 8-byte aligned.
The existing bug is that we are not accounting for this extra space that
will be required, when determining whether we will have enough space to
store a container.
This resulted in rare cases in creating a segment size that exceeded the
size of a block, thankfully causing an InternalException.

The other bug is in the RoaringBoolean's ScanPartial / FetchRow
implementations. It was reading at an offset that was out of bounds of
the dummy vector that was created.
And the scanned data was written to the result Vector at a wrong offset.

Thankfully none of these were bugs in the compression logic, so no
borked files were created as a result of these bugs.

Upstream-commit: https://github.com/duckdb/duckdb/commit/d2cbbf369aed91cd39e30e9817faea279cdd84f2