PEG Transformer template script (https://redirect.github.com/duckdb/duckdb/pull/19812)
Part of epic: https://github.com/duckdblabs/duckdb-internal/issues/6380

Hi there,
This PR introduces a script `generate_peg_transformer.py` that looks at
all the grammar files in `autocomplete/statement` and tries to find the
corresponding transformer rule.
It outputs per `*.gram` file which rules are still missing a transformer
rule. In most cases, every grammar rule corresponds to a transformer
rule for the PEG parser. Rules can be excluded if we don't require a
transformer rule.
For `alter.gram` for instance, the output will as follows to indicate
that these transformer rules are still missing:
```

Upstream-commit: https://github.com/duckdb/duckdb/commit/2ca8f2015dd48a352f0226e3888e1ffe7798d3ef
--- File: alter.gram ---
[ MISSING ]    AlterSchemaStmt
[ MISSING ]    SetData
``` 

Additionally, when using the `-g` argument, the script prints a template
for every missing transformer rule (in alphabetical order). It will
output three parts:
1. The method declaration that should be pasted into
`peg_transformer.hpp`
2. The method implementation 
3. The transformer rule registation
The standard type that is used is `unique_ptr<SQLStatement>`. This is
most likely not correct, but at the moment it is not possible to know
the type in this template script. We could add a file where we can
specify the return type for every rule if this is important enough. For
now I added a `// TODO` to remind users to check the return type.

I decided against letting this script modify the `transform_*.cpp` files
directly to avoid making the script overly complex. Perhaps in a future
update this could be looked into.

For example, for `WithinGroupClause`: 
```
--- Generation for rule: WithinGroupClause ---
1. Add DECLARATION to: ../extension/autocomplete/include/transformer/peg_transformer.hpp
// TODO: Verify this return type is correct
    static unique_ptr<SQLStatement> TransformWithinGroupClause(PEGTransformer &transformer, optional_ptr<ParseResult> parse_result);

3. Add REGISTRATION to: ../extension/autocomplete/transformer/peg_transformer_factory.cpp
Inside the appropriate Register...() function:
REGISTER_TRANSFORM(TransformWithinGroupClause);

4. Add IMPLEMENTATION to: ../extension/autocomplete/transformer/transform_expression.cpp
// TODO: Verify this return type is correct
unique_ptr<SQLStatement> PEGTransformerFactory::TransformWithinGroupClause(PEGTransformer &transformer,
                                                                   optional_ptr<ParseResult> parse_result) {
	throw NotImplementedException("TransformWithinGroupClause has not yet been implemented");
}

--- End of WithinGroupClause ---
```

In a future PR, when the transformer is close to completion, I would
like to use this file in a workflow to check for missing transformer
rules.

Finally, it will output orphan rules, transformer rules that don't have
a corresponding grammar rule. To avoid keeping transformer rules around
that don't do anything.

There are currently two arguments: 
- `-g` or `--generate`: 
- `-s` or `--skip-found`: Skips the rules that have the status [ FOUND ]
or [ ENUM ] to keep the output a bit cleaner.