[Testing][Dev] Extend tests covered by the `storage_compatibility` test config (https://redirect.github.com/duckdb/duckdb/pull/19860)
Also replace the `load` paths when we're running the
`storage_compatibility` test config.

As can be seen by the updated list of skipped tests, we now also
properly test the Roaring boolean tests, and have to skip them because
they are (correctly) not forwards-compatible.

I've had to add a new list of skipped tests because of various problems
with doing this path replacement for `load`, such as:
- `load` can be used multiple times in a test, with different paths
- The paths used in `load` can be referenced by the test (detach,
attach, export, import, etc..)

Parts of these could be fixed by turning the loaded path into a
variable, so we could reference `${ACTIVE_DATABASE_PATH}` for example.

### Misc changes

Fixed an uncaught error if the provided version isn't valid (the
download fails)
Print the errors to stderr rather than stdout, for easier filtering of
the script output.
Added prints for various reasons that a test was already being skipped
without being marked as `[SKIPPED]`, for better debugging of the script.

Upstream-commit: https://github.com/duckdb/duckdb/commit/f69183bb1a87127af9ba9dfb6f1674f89b69145b