[Parquet] Implement automatic shredding for `VARIANT` columns in COPY TO parquet (https://redirect.github.com/duckdb/duckdb/pull/19336)
This PR is a follow up to https://redirect.github.com/duckdb/duckdb/pull/19219 

Which added the ability to provide a `SHREDDING` copy option to manually
set the shredded type of a VARIANT column.
This implemented shredding and made it easy to control, but it requires
manual configuration by the user.

With this PR, VARIANT columns will always be shredded based on the
analysis of the first rowgroup.
(TODO: this should be queryable with `parquet_metadata` /
`parquet_schema`)

This required some changes to the Parquet writing path:
- `SchemaElement` items are no longer created during bind.
- `ParquetColumnSchema` `schema_index` is now an `optional_idx`, as this
index refers to the `SchemaElements` vector, which is populated later.
- `ColumnWriter` no longer takes a reference to a `ParquetColumnSchema`
object, instead it now owns the schema.
- `FillParquetSchema` is removed, consolidated into
`CreateWriterRecursive`
- [1]`ParquetWriteTransformData` is introduced, this struct holds the
reusable state needed to perform a pre-processing step to transform the
input columns into the shape required by the ColumnWriter (relevant for
VARIANT).
- `FinalizeSchema` has been added to create the `SchemaElements`, and
populate the `schema_index` in the schema objects.
- `HasTransform`, `TransformedType` and `TransformExpression` are added
to perform the pre-processing step, using the
`ParquetWriteTransformData`.
- `AnalyzeSchemaInit`, `AnalyzeSchema` and `AnalyzeSchemaFinalize` are
added to auto-detect the shredding for VARIANT columns and apply this
change to the schema, to be used before `FinalizeSchema` is called.

[1] This is not as clean as I want it to be, the
`ParquetWriteTransformData` can't only live in the
`ParquetWriteLocalState`, there also needs to be a copy in the
`ParquetWriteGlobalState` as well as the `ParquetWritePrepareBatch`
method.

### Limitations

`DECIMAL` types are never automatically shredded on.
This is because the DECIMAL type is somewhat special, in that not only
the underlying type has to be the same, the width+scale also has to be
the same, so I've opted to not collect data for that.

Upstream-commit: https://github.com/duckdb/duckdb/commit/9b445d1a25493b2073e56249fcd37017ed4a4c57